refactor: extract layout bootstrap
